Corrected package.json file and added gitignore file
