+++ database_models/UserSchema.js
@@ -19,6 +19,7 @@ const UserSchema = new Schema({
         default: Date.now
     },
 
-});
+}, { timestamps: true }
+);
 const User = mongoose.model('user', UserSchema);
 module.exports = User;
