+++ index.js
@@ -1,7 +1,7 @@
 const connectTodatabase = require('./database');
 const express = require('express')
 const app = express()
-const port = 5000
+const port = process.env.PORT || 5000
 
 connectTodatabase();
