+++ package.json
@@ -4,7 +4,8 @@
   "description": "backend of diary app",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index",
+    "dev": "nodemon index"
   },
   "author": "divi osho",
   "license": "ISC",
